refactor(rich_text_editor): extract tests to separate module
- Extract 1,146 lines of tests (140 test functions) to tests.rs
- Reduces mod.rs from 1,301 to 156 lines
- Completes the monolithic file split into 18 focused modules
